refactor: use NumberFormat

+++ packages/code-du-travail-frontend/src/outils/CommonIndemniteDepart/steps/Resultat/components/FilledElements.tsx
@@ -149,7 +149,7 @@ export default function FilledElements(props: Props) {
                   inclus)&nbsp;:&nbsp;
                   <span
                     dangerouslySetInnerHTML={{
-                      __html: formatNumber(parseInt(props.salary)),
+                      __html: formatNumber(parseFloat(props.salary)),
                     }}
                   />
                 </li>
